* version bump to v0.2.2-b1, time to call this a beta agin  * hard-disabled vulkan support in build.sh (this is very incomplete at this point and should not be used - this will come back with v0.3)  * added new math constants, improved existing ones, changed some names for consistency  * added a config option to disable spir-v support (and use spir instead), on opencl 2.1+ platforms or those supporting cl_khr_il_program spir-v is chosen over spir, but spir-v support is still rather experimental at this point (-> allow to disable it)  * fixed clang 3.5 compilation (added empty __builtin_assume define, since __builtin_assume is not supported in 3.5)  * made kernel max_local_work_size an uint32_t, because that is more than enough and allows it to be directly used for setting global/local work dims

+++ compute/compute_kernel.hpp
@@ -38,7 +38,7 @@ public:
 	
 	struct kernel_entry {
 		const llvm_compute::function_info* info { nullptr };
-		size_t max_local_work_size { 0u };
+		uint32_t max_local_work_size { 0u };
 		uint3 max_work_group_item_sizes;
 	};
 	//! returns the internal kernel entry for the specified device
